Convert ProfileCOntainer to Typescript and use new components
